requirements.txt
asgiref==3.6.0
attrs==22.2.0
Django==4.2
django-js-asset==2.0.0
django-mptt==0.14.0
djangorestframework==3.14.0
drf-spectacular==0.26.1
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.17.3
packaging==23.1
pluggy==1.0.0
pyrsistent==0.19.3
pytest==7.3.0
pytest-django==4.5.2
python-decouple==3.8
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.3
uritemplate==4.1.1
